sweet-harmony-manager/Backend/requirements.txt
django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.8.2
supabase==2.6.0